+++ config/application.pantheon.php
@@ -49,9 +49,3 @@ if ( isset( $_ENV['PANTHEON_ENVIRONMENT'] ) ) {
 		define( 'PANTHEON_HOSTNAME', $hostname );
 	}
 }
-
-// Update the multisite configuration to use Config::define() instead of define.
-add_filter( 'pantheon.multisite.config_contents', function ( $config_contents ) {
-	$config_contents = str_replace( 'define(', 'Config::define(', $config_contents );
-	return $config_contents;
-} );